==> 04_Analyze Spain First Author Publications.ipynb <==
# Analyze Spain First Author Publications

#### Updated: Sep 22, 2022

#  

Analyze subset of publications with Spanish first authors. Use this subset to identify top Spanish researchers related to concepts of interest to GAC-MDRID.

import pandas as pd
import numpy as np
import os
import ast
import time

dataDir = '/Users/rnaidoo/Documents/Canada-Secure/GAC/2022_MDRID/Projects_data/OpenAlex/works_esp_first_auth/'

#  

#### Functions:

def search_concepts(search_query, df_concepts):
    
    search_terms = search_query.split(' AND ')
    for i in range(0, len(search_terms)):
        search_terms[i] = search_terms[i].lower()
    print(search_terms)
    for i in range(0, len(search_terms)):
        if i == 0:
            df_search = df_concepts.loc[df_concepts['concept'].str.contains(search_terms[i])]
        else:
            df_search_ = df_concepts.loc[df_concepts['concept'].str.contains(search_terms[i])]
            df_search_ = df_search_.rename(columns={
                'concept': 'concept'+str(i+1),
                'concept_id': 'concept_id'+str(i+1),
                'concept_level': 'concept_level'+str(i+1)
            })
            df_search = df_search.merge(df_search_, how='inner', on='pub_id')
    print(str(len(df_search)) + ' papers related to search...')
    
    return df_search

def search_title(search_query, df_pubs, keep_case=False):
    
    search_terms = search_query.split(' AND ')
    if not keep_case:
        for i in range(0, len(search_terms)):
            search_terms[i] = search_terms[i].lower()
        df_pubs['title'] = df_pubs['title'].str.lower()
    print(search_terms)
    for i in range(0, len(search_terms)):
        if i == 0:
            df_search = df_pubs.loc[df_pubs['title'].str.contains(search_terms[i])]
        else:
            df_search = df_search.loc[df_search['title'].str.contains(search_terms[i])]
            #df_search = pd.concat([df_search, df_search_])
    print(str(len(df_search)) + ' papers related to search...')
    
    return df_search

def retrieve_papers_byConcept(search_query, df_search, df_pubs):
    
    df_merge = df_pubs.merge(df_search, left_on='id', right_on='pub_id')
    print(str(len(df_merge)) + ' papers retrieved...')
    df_result = df_merge[['first_author', 'first_auth_inst', 'concept', 'concept_level', 'title', 'publication_date', 'cited_by_count', 'collaborators_of_interest', 'first_auth_orcid', 'id']]
    df_result = df_result.sort_values(['cited_by_count', 'publication_date', 'concept_level'], ascending=[False, False, False]).reset_index(drop=True)
    df_result = df_result.drop_duplicates(subset=['id'], keep='first')
    print(str(len(df_result)) + ' unique papers retrieved.')
    df_result = df_result.drop(columns=['concept_level'])
    df_result = df_result.rename(columns={
        'first_author': 'Spanish Researcher',
        'first_auth_inst': 'Institution',
        'concept': 'Concept',
        'title': 'Publication Title',
        'publication_date': 'Publication Date',
        'cited_by_count': 'Citation Count',
        'collaborators_of_interest': 'Canadian Collaborators',
        'first_auth_orcid': 'Academic Profile (ORCID)',
        'id': 'Publication Profile'
    })
    df_result = df_result.replace(np.nan, '')
    df_result['Canadian Collaborators'] = df_result['Canadian Collaborators'].replace('{}', 'N')

    #Process Canadian Collaborators
    for i in range(0, len(df_result)):
        collab_dict_ = df_result['Canadian Collaborators'].iloc[i]
        if collab_dict_ != 'N':
            collab_dict = ast.literal_eval(collab_dict_)
            if len(collab_dict) > 0:
                collab_str = ''
                for auth in collab_dict:
                    if collab_str != '':
                        collab_str += ', '
                    collab_str += auth + ' (' + collab_dict[auth]['col_auth_inst']
                    if collab_dict[auth]['col_auth_orcid'] == None:
                        collab_str += ')'
                    else: 
                        collab_str += ', ' + collab_dict[auth]['col_auth_orcid'] + ')'
                df_result.iloc[i,6] = collab_str
                
    return df_result        

def retrieve_papers_byTitle(search_query, df_search):
    
    print(str(len(df_search)) + ' papers retrieved...')
    df_result = df_search[['first_author', 'first_auth_inst', 'title', 'publication_date', 'cited_by_count', 'collaborators_of_interest', 'first_auth_orcid', 'id']]
    df_result = df_result.sort_values(['cited_by_count', 'publication_date'], ascending=[False, False]).reset_index(drop=True)
    df_result = df_result.drop_duplicates(subset=['id'], keep='first')
    print(str(len(df_result)) + ' unique papers retrieved.')
    df_result = df_result.rename(columns={
        'first_author': 'Spanish Researcher',
        'first_auth_inst': 'Institution',
        'title': 'Publication Title',
        'publication_date': 'Publication Date',
        'cited_by_count': 'Citation Count',
        'collaborators_of_interest': 'Canadian Collaborators',
        'first_auth_orcid': 'Academic Profile (ORCID)',
        'id': 'Publication Profile'
    })
    df_result = df_result.replace(np.nan, '')
    df_result['Canadian Collaborators'] = df_result['Canadian Collaborators'].replace('{}', 'N')
    
    #Process Canadian Collaborators
    for i in range(0, len(df_result)):
        collab_dict_ = df_result['Canadian Collaborators'].iloc[i]
        if collab_dict_ != 'N':
            collab_dict = ast.literal_eval(collab_dict_)
            if len(collab_dict) > 0:
                collab_str = ''
                for auth in collab_dict:
                    if collab_str != '':
                        collab_str += ', '
                    collab_str += auth + ' (' + collab_dict[auth]['col_auth_inst']
                    if collab_dict[auth]['col_auth_orcid'] == None:
                        collab_str += ')'
                    else: 
                        collab_str += ', ' + collab_dict[auth]['col_auth_orcid'] + ')'
                df_result.iloc[i,5] = collab_str
                
    return df_result        

def write_result(search_query, df_result, dataDir_save, search_by_title=False):
    
    sheet_name_ = search_query[0:30]
    if search_by_title:
        sbt_str = 'search_title '
    else:
        sbt_str = ''
    writer = pd.ExcelWriter(dataDir_save + 'works_by_search/' + 'works_by_' + sbt_str + search_query + '.xlsx')
    df_result.to_excel(writer, sheet_name=sheet_name_, index=False)
    # Auto-adjust columns' width
    for column in df_result:
        if column == 'Publication Title':
            column_width = 100
        elif column == 'Canadian Collaborators':
            if len(df_result['Canadian Collaborators'].unique()) > 1:
                column_width = 75
            else:
                column_width = max(df_result[column].astype(str).map(len).max(), len(column))
        else:
            column_width = max(df_result[column].astype(str).map(len).max(), len(column))
        col_idx = df_result.columns.get_loc(column)
        writer.sheets[sheet_name_].set_column(col_idx, col_idx, column_width)

    writer.save()

#  

#### Combine yearly batches, remove duplicate publications, split by year again:

df_total = pd.DataFrame()
for file in os.listdir(dataDir):
    if 'works_esp_first_auth_20' in file:
        if '_proc' in file:
            #print(file)
            df_load = pd.read_csv(dataDir + file)
            print('Opening ' + file + '...rows: ' + str(len(df_load)))
            df_total = pd.concat([df_total, df_load])

df_total

df_total = df_total.sort_values('publication_date')
df_total.tail()

df1 = df_total.drop_duplicates(subset=['id'])
df1

df1.to_csv(dataDir + 'works_esp_first_auth_since2017.csv', index=False)

df1 = pd.read_csv(dataDir + 'works_esp_first_auth_since2017.csv')
df1

pub_years = df1['publication_year'].unique()
pub_years

for year in pub_years:
    df1_y = df1.loc[df1['publication_year'] == year]
    df1_y.to_csv(dataDir + 'works_esp_first_auth_' + str(year) + '_proc2.csv', index=False)

#  

#### Process data to surface concepts related to each publication, by year:

Manually process each year:

year = 2022
df_pubs = pd.read_csv(dataDir + 'works_esp_first_auth_' + str(year) + '_proc2.csv')
print(len(df_pubs))
df_pubs.head()

batch_size = 10000

start = time.time()
ind = 0
for i in range(0, len(df_pubs)):
    if i == 0:
        start_batch = time.time()
        df_concepts_batch = pd.DataFrame()
    
    pub_id = df_pubs.iloc[i]['id']
    #Get concepts
    concepts_ = df_pubs.iloc[i]['concepts']
    concepts = ast.literal_eval(concepts_)
    for j in range(0, len(concepts)):
        ind+=1
        row = pd.DataFrame(data={
            'pub_id': pub_id,
            'concept': concepts[j]['display_name'],
            'concept_id': concepts[j]['id'],
            'concept_level': concepts[j]['level']
        }, index=[ind])
        df_concepts_batch = pd.concat([df_concepts_batch, row]) 
    
    if (np.remainder(i, batch_size) == 0) & (i>0):
        if i == batch_size:
            df_concepts_batch.to_csv(dataDir + 'concepts_by_pub_esp_first_auth_' + str(year) + '.csv', index=False)
        else:
            df_concepts = pd.read_csv(dataDir + 'concepts_by_pub_esp_first_auth_' + str(year) + '.csv')
            df_concepts = pd.concat([df_concepts, df_concepts_batch])
            df_concepts.to_csv(dataDir + 'concepts_by_pub_esp_first_auth_' + str(year) + '.csv', index=False)
            print(str(len(df_concepts)) + ' rows saved.')
        end_batch = time.time()
        t_batch = end_batch - start_batch
        rate_batch = batch_size/t_batch
        t_remaining = round(((len(df_pubs) - i)/rate_batch)/60,0)
        print('Processing row ' + str(i) + '...at rate ' + str(round(rate_batch,0)) + ' rows/s...' + str(t_remaining) + ' min estimated remaining.')
        start_batch = time.time()
        df_concepts_batch = pd.DataFrame()
        
end = time.time()   
t = end - start
print('Runtime: ' + str(round(t/60, 1)) + ' min')

df_concepts = pd.read_csv(dataDir + 'concepts_by_pub_esp_first_auth_' + str(year) + '.csv')
df_concepts['concept'] = df_concepts['concept'].str.lower()
df_concepts

df_concepts.to_csv(dataDir + 'concepts_by_pub_esp_first_auth_' + str(year) + '.csv', index=False)

Combine yearly batches

df_total = pd.DataFrame()
for file in os.listdir(dataDir):
    if 'concepts_by_pub_esp_first_auth_' in file:
        #print(file)
        df_load = pd.read_csv(dataDir + file)
        print('Opening ' + file + '...rows: ' + str(len(df_load)))
        df_total = pd.concat([df_total, df_load])

df_total

df_total = df_total.sort_values('concept_level').reset_index(drop=True)
df_total.tail()

df_total.to_csv(dataDir + 'concepts_by_pub_esp_first_auth_since2017.csv', index=False)

#  

#### Identify top Spanish researchers related to concepts of interest to GAC-MDRID:

Ref: evernote:///view/10251849/s92/c2b57ab9-cbf3-423c-86e7-5e2094ce32a3/8137299e-1b5c-c964-fcc6-1a32ffc0d0f9

df_pubs = pd.read_csv(dataDir + 'works_esp_first_auth_since2017.csv')
print(len(df_pubs))
df_pubs.head()

df_concepts = pd.read_csv(dataDir + 'concepts_by_pub_esp_first_auth_since2017.csv')
df_concepts

'federalism'

search_query = 'cities'

df_search = search_concepts(search_query=search_query, df_concepts=df_concepts)
df_result = retrieve_papers(search_query=search_query, df_search=df_search, df_pubs=df_pubs)
write_result(search_query=search_query, df_result=df_result, dataDir_save=(dataDir))
df_result.head()

#  

For a series of search queries:

search_queries = [
    'European Union',
    'Ukraine',
]

for search_query in search_queries:
    df_search = search_concepts(search_query=search_query, df_concepts=df_concepts)
    df_result = retrieve_papers(search_query=search_query, df_search=df_search, df_pubs=df_pubs)
    write_result(search_query=search_query, df_result=df_result)
    print('')

#  

#### Search keywords by title:

df_pubs2 = df_pubs.fillna('')
df_pubs2.head()

search_query = 'Canada-EU Free Trade Agreement'

df_search = search_title(search_query=search_query, df_pubs=df_pubs2, keep_case=False)
df_result = retrieve_papers_byTitle(search_query=search_query, df_search=df_search)
write_result(search_query=search_query, df_result=df_result, dataDir_save=(dataDir), search_by_title=True)
df_result.head()
