==> 05_Regional_Profile.ipynb <==
# Regional Profile

#### Updated: Sep 15, 2022

#  

The Ambassador will make a visit to the University of Granada in December 2022. Prepare a profile of the university, to surface the top academics overall, the top academics for particular subject-areas, and the top collaborators with Canada. Generalize the production of regional profiles:

import pandas as pd
import numpy as np
import ast

dataDir = '/Users/rnaidoo/Documents/Canada-Secure/GAC/2022_MDRID/Projects_data/OpenAlex/'

#  

#### Functions:

def search_institutions(df_pubs, inst_search_terms=[]):
    
    inst_search_results = []
    for search_term in inst_search_terms:
        inst_search_results = inst_search_results + list(df_pubs.loc[df_pubs['first_auth_inst'].str.contains(search_term)]['first_auth_inst'].unique())
    df_insts = pd.DataFrame({'inst_search': inst_search_results})
    
    df_pubs2 = df_pubs.merge(df_insts, how='inner', left_on='first_auth_inst', right_on='inst_search')
    
    return [df_insts, df_pubs2]

def top_institutions(df_pubs, reg_name, top=100):
    
    df_pubs['publications'] = 1
    df_top_inst = df_pubs[['cited_by_count', 'publications', 'first_auth_inst']].groupby(['first_auth_inst']).sum()
    df_top_inst = df_top_inst.sort_values('cited_by_count', ascending=False)
    df_top_inst = df_top_inst.replace(np.nan, '')
    df_top_inst = df_top_inst.rename(columns={
        'cited_by_count': 'by citations',
        'publications': 'by publications'
    })
    if len(df_top_inst) < top:
        top = len(df_top_inst)
    df_top_inst.index.names = ['Top-' + str(top) + ' research institutions in ' + reg_name]
    
    return df_top_inst.head(top)

def top_researchers(df_pubs, reg_name, top=100):
    
    df_pubs['publications'] = 1
    df_top_ac = df_pubs[['cited_by_count', 'publications', 'first_author']].groupby(['first_author']).sum()
    df_top_ac = df_top_ac.sort_values('cited_by_count', ascending=False)
    df_top_ac2 = df_top_ac.merge(df_pubs[['first_author', 'first_auth_inst', 'first_auth_orcid']], how='left', on='first_author')
    df_top_ac2 = df_top_ac2.replace(np.nan, '')
    df_top_ac2 = df_top_ac2.rename(columns={
        'cited_by_count': 'by citations',
        'publications': 'by publications',
        'first_auth_inst': 'institution',
        'first_auth_orcid': 'orcid'
    })
    df_top_ac2 = df_top_ac2.drop_duplicates(subset=['orcid'])
    df_top_ac2 = df_top_ac2.drop_duplicates(subset=['first_author'])
    df_top_ac2 = df_top_ac2.set_index('first_author')
    if len(df_top_ac2) < top:
        top = len(df_top_ac2)
    df_top_ac2.index.names = ['Top-' + str(top) + ' academics in ' + reg_name]
    
    return df_top_ac2.head(top)

def search_concepts(search_query, df_concepts):
    
    search_terms = search_query.split(' AND ')
    for i in range(0, len(search_terms)):
        search_terms[i] = search_terms[i].lower()
    print(search_terms)
    for i in range(0, len(search_terms)):
        if i == 0:
            df_search = df_concepts.loc[df_concepts['concept'].str.contains(search_terms[i])]
        else:
            df_search_ = df_concepts.loc[df_concepts['concept'].str.contains(search_terms[i])]
            df_search_ = df_search_.rename(columns={
                'concept': 'concept'+str(i+1),
                'concept_id': 'concept_id'+str(i+1),
                'concept_level': 'concept_level'+str(i+1)
            })
            df_search = df_search.merge(df_search_, how='inner', on='pub_id')
    print(str(len(df_search)) + ' papers related to search...')
    
    return df_search

def write_pubs_by_concept_search(search_query, df_result, dataDir_save, fn_prefix=''):
    
    sheet_name_ = search_query[0:30]
    writer = pd.ExcelWriter(dataDir_save + fn_prefix + 'works_by_' + search_query + '.xlsx')
    df_result.to_excel(writer, sheet_name=sheet_name_, index=False)
    # Auto-adjust columns' width
    for column in df_result:
        if column == 'Publication Title':
            column_width = 100
        elif column == 'Canadian Collaborators':
            if len(df_result['Canadian Collaborators'].unique()) > 1:
                column_width = 75
            else:
                column_width = max(df_result[column].astype(str).map(len).max(), len(column))
        else:
            column_width = max(df_result[column].astype(str).map(len).max(), len(column))
        col_idx = df_result.columns.get_loc(column)
        writer.sheets[sheet_name_].set_column(col_idx, col_idx, column_width)

    writer.save()

def top_researchers_by_concept(search_query, df_pubs, df_concepts, reg_name, dataDir_save, top=100):

    df_search = search_concepts(search_query=search_query, df_concepts=df_concepts)
    df_pubs_search = df_pubs.merge(df_search, left_on='id', right_on='pub_id')
    print(str(len(df_pubs_search)) + ' papers retrieved...')
    
    df_result = df_pubs_search[['first_author', 'first_auth_inst', 'concept', 'concept_level', 'title', 'publication_date', 'cited_by_count', 'collaborators_of_interest', 'first_auth_orcid', 'id']]
    df_result = df_result.sort_values(['cited_by_count', 'publication_date', 'concept_level'], ascending=[False, False, False]).reset_index(drop=True)
    df_result = df_result.drop_duplicates(subset=['id'], keep='first')
    print(str(len(df_result)) + ' unique papers retrieved.')
    df_result = df_result.drop(columns=['concept_level'])
    df_result = df_result.rename(columns={
        'first_author': 'Spanish Researcher',
        'first_auth_inst': 'Institution',
        'concept': 'Concept',
        'title': 'Publication Title',
        'publication_date': 'Publication Date',
        'cited_by_count': 'Citation Count',
        'collaborators_of_interest': 'Canadian Collaborators',
        'first_auth_orcid': 'Academic Profile (ORCID)',
        'id': 'Publication Profile'
    })
    df_result = df_result.replace(np.nan, '')
    df_result['Canadian Collaborators'] = df_result['Canadian Collaborators'].replace('{}', 'N')

    #Process Canadian Collaborators
    for i in range(0, len(df_result)):
        collab_dict_ = df_result['Canadian Collaborators'].iloc[i]
        if collab_dict_ != 'N':
            collab_dict = ast.literal_eval(collab_dict_)
            if len(collab_dict) > 0:
                collab_str = ''
                for auth in collab_dict:
                    if collab_str != '':
                        collab_str += ', '
                    collab_str += auth + ' (' + collab_dict[auth]['col_auth_inst']
                    if collab_dict[auth]['col_auth_orcid'] == None:
                        collab_str += ')'
                    else: 
                        collab_str += ', ' + collab_dict[auth]['col_auth_orcid'] + ')'
                df_result.iloc[i,6] = collab_str
                
    #Write results to Excel spreadsheet
    write_pubs_by_concept_search(search_query=search_query, df_result=df_result, dataDir_save=dataDir_save, fn_prefix=reg_name+'_')
    
    df_pubs_search['publications'] = 1
    df_top_ac = df_pubs_search[['cited_by_count', 'publications', 'first_author']].groupby(['first_author']).sum()
    df_top_ac = df_top_ac.sort_values('cited_by_count', ascending=False)
    df_top_ac2 = df_top_ac.merge(df_pubs[['first_author', 'first_auth_inst', 'first_auth_orcid']], how='left', on='first_author')
    df_top_ac2 = df_top_ac2.replace(np.nan, '')
    df_top_ac2 = df_top_ac2.rename(columns={
        'cited_by_count': 'by citations',
        'publications': 'by publications',
        'first_auth_inst': 'institution',
        'first_auth_orcid': 'orcid'
    })
    df_top_ac2 = df_top_ac2.drop_duplicates(subset=['orcid'])
    df_top_ac2 = df_top_ac2.drop_duplicates(subset=['first_author'])
    df_top_ac2 = df_top_ac2.set_index('first_author')
    if len(df_top_ac2) < top:
        top = len(df_top_ac2)
    df_top_ac2.index.names = ['Top-' + str(top) + ' academics in ' + reg_name + ' for ' + search_query]
    
    return df_top_ac2.head(top)

def write_pubs_by_top_collab_Can(df_result, dataDir_save, fn_prefix=''):

    sheet_name_ = 'top-pubs with Canadian collab'
    writer = pd.ExcelWriter(dataDir_save + fn_prefix + 'works_can_col.xlsx')
    df_result.to_excel(writer, sheet_name=sheet_name_, index=False)
    # Auto-adjust columns' width
    for column in df_result:
        if column == 'Publication Title':
            column_width = 100
        elif column == 'Canadian Collaborators':
            column_width = 75
        else:
            column_width = max(df_result[column].astype(str).map(len).max(), len(column))
        col_idx = df_result.columns.get_loc(column)
        writer.sheets[sheet_name_].set_column(col_idx, col_idx, column_width)

    writer.save()

def top_collab_Can(df_pubs, reg_name, dataDir_save, top=100):
    
    df_can_col = df_pubs.loc[df_pubs['collaborators_of_interest'] != '{}']
    
    df_result = df_can_col[['first_author', 'first_auth_inst', 'title', 'publication_date', 'cited_by_count', 'collaborators_of_interest', 'first_auth_orcid', 'id']]
    df_result = df_result.sort_values(['cited_by_count', 'publication_date'], ascending=[False, False]).reset_index(drop=True)
    df_result = df_result.drop_duplicates(subset=['id'], keep='first')
    print(str(len(df_result)) + ' unique papers retrieved.')
    df_result = df_result.rename(columns={
        'first_author': 'Spanish Researcher',
        'first_auth_inst': 'Institution',
        'title': 'Publication Title',
        'publication_date': 'Publication Date',
        'cited_by_count': 'Citation Count',
        'collaborators_of_interest': 'Canadian Collaborators',
        'first_auth_orcid': 'Academic Profile (ORCID)',
        'id': 'Publication Profile'
    })
    df_result = df_result.replace(np.nan, '')

    #Process Canadian Collaborators
    for i in range(0, len(df_result)):
        collab_dict_ = df_result['Canadian Collaborators'].iloc[i]
        if collab_dict_ != 'N':
            collab_dict = ast.literal_eval(collab_dict_)
            if len(collab_dict) > 0:
                collab_str = ''
                for auth in collab_dict:
                    if collab_str != '':
                        collab_str += ', '
                    collab_str += auth + ' (' + collab_dict[auth]['col_auth_inst']
                    if collab_dict[auth]['col_auth_orcid'] == None:
                        collab_str += ')'
                    else: 
                        collab_str += ', ' + collab_dict[auth]['col_auth_orcid'] + ')'
                df_result.iloc[i,5] = collab_str
    
    #Write results to Excel spreadsheet
    write_pubs_by_top_collab_Can(df_result=df_result, dataDir_save=dataDir_save, fn_prefix=reg_name+'_')
    
    df_can_col['publications'] = 1
    df_top_ac = df_can_col[['cited_by_count', 'publications', 'first_author']].groupby(['first_author']).sum()
    df_top_ac = df_top_ac.sort_values('cited_by_count', ascending=False)
    df_top_ac2 = df_top_ac.merge(df_pubs[['first_author', 'first_auth_inst', 'first_auth_orcid']], how='left', on='first_author')
    df_top_ac2 = df_top_ac2.replace(np.nan, '')
    df_top_ac2 = df_top_ac2.rename(columns={
        'cited_by_count': 'by citations',
        'publications': 'by publications',
        'first_auth_inst': 'institution',
        'first_auth_orcid': 'orcid'
    })
    df_top_ac2 = df_top_ac2.drop_duplicates(subset=['orcid'])
    df_top_ac2 = df_top_ac2.drop_duplicates(subset=['first_author'])
    df_top_ac2 = df_top_ac2.set_index('first_author')
    if len(df_top_ac2) < top:
        top = len(df_top_ac2)
    df_top_ac2.index.names = ['Top-' + str(top) + ' academics in ' + reg_name + ' that collaborate with Canadian researchers']
    
    return df_top_ac2.head(top)

#  

#### Load requisite data:

df_pubs = pd.read_csv(dataDir + 'works_esp_first_auth/' + 'works_esp_first_auth_since2017.csv')
print(len(df_pubs))
df_pubs.head()

df_concepts = pd.read_csv(dataDir + 'works_esp_first_auth/' + 'concepts_by_pub_esp_first_auth_since2017.csv')
df_concepts

#  

#### Analyze a region of interest:

reg_name = 'Madrid'

dataDir_reg = dataDir+'works_esp_first_auth/Regional/'+reg_name+'/'

inst_search_terms = [
    
    #Madrid
    'Madrid',
    'IE',
    'Consejo Superior',
    'Spanish National Research Council'
    
    #Barcelona
    #'Barc',
    #'Cata'
    
    #Granada
    #'Granada',
    #'Andalu'
]

df_insts, df_pubs_reg = search_institutions(df_pubs=df_pubs, inst_search_terms=inst_search_terms)
df_insts['inst_search'].unique()

top_institutions(df_pubs=df_pubs_reg, reg_name=reg_name, top=10)

top_researchers(df_pubs=df_pubs_reg, reg_name=reg_name, top=10)

top_collab_Can(df_pubs=df_pubs_reg, reg_name=reg_name, dataDir_save=dataDir_reg, top=10)

search_query = 'foreign policy'
top_researchers_by_concept(search_query=search_query, df_pubs=df_pubs_reg, df_concepts=df_concepts, reg_name=reg_name, dataDir_save=dataDir_reg, top=10)

search_query = 'international relations'
top_researchers_by_concept(search_query=search_query, df_pubs=df_pubs_reg, df_concepts=df_concepts, reg_name=reg_name, dataDir_save=dataDir_reg, top=10)

search_query = 'climate change'
top_researchers_by_concept(search_query=search_query, df_pubs=df_pubs_reg, df_concepts=df_concepts, reg_name=reg_name, dataDir_save=dataDir_reg, top=10)

#  

## Development

#### Determine all of the institutions in Granada:

df_pubs = pd.read_csv(dataDir + 'works_esp_first_auth/' + 'works_esp_first_auth_since2017.csv')
df_pubs.head()

inst_search_terms = [
    'Granada',
    'Andalu'
]

inst_search_results = []
for search_term in inst_search_terms:
    inst_search_results = inst_search_results + list(df_pubs.loc[df_pubs['first_auth_inst'].str.contains(search_term)]['first_auth_inst'].unique())

df_insts = pd.DataFrame({'inst_search': inst_search_results})
df_insts

df_pubs2 = df_pubs.merge(df_insts, how='inner', left_on='first_auth_inst', right_on='inst_search')
print(len(df_pubs2))
df_pubs2.head()

Top regional institutions, overall:

df_pubs2['publications'] = 1

df_top_inst = df_pubs2[['cited_by_count', 'publications', 'first_auth_inst']].groupby(['first_auth_inst']).sum()
df_top_inst = df_top_inst.sort_values('cited_by_count', ascending=False)
df_top_inst = df_top_inst.rename(columns={
    'cited_by_count': 'by citations',
    'publications': 'by publications'
})
df_top_inst.index.names = ['top research institutions in Andalucía']
df_top_inst.head(10)

Top regional researchers, overall:

df_top_ac = df_pubs2[['cited_by_count', 'publications', 'first_author']].groupby(['first_author']).sum()
df_top_ac = df_top_ac.sort_values('cited_by_count', ascending=False)
df_top_ac2 = df_top_ac.merge(df_pubs2[['first_author', 'first_auth_inst', 'first_auth_orcid']], how='left', on='first_author')
df_top_ac2 = df_top_ac2.rename(columns={
    'cited_by_count': 'by citations',
    'publications': 'by publications',
    'first_auth_inst': 'institution',
    'first_auth_orcid': 'orcid'
})
df_top_ac2 = df_top_ac2.drop_duplicates(subset=['orcid'])
df_top_ac2 = df_top_ac2.set_index('first_author')
df_top_ac2.index.names = ['Top-20 academics in Andalucía']
df_top_ac2.head(20)

Top regional researchers in a particular subject area:

df_concepts = pd.read_csv(dataDir + 'works_esp_first_auth/' + 'concepts_by_pub_esp_first_auth_since2017.csv')
df_concepts

'climate change'

search_query = 'climate change'

dataDir_save = dataDir + 'works_esp_first_auth/Regional/Granada/'
df_search = search_concepts(search_query=search_query, df_concepts=df_concepts)
df_result = retrieve_papers(search_query=search_query, df_search=df_search, df_pubs=df_pubs2)
write_result(search_query=search_query, df_result=df_result, dataDir_save=dataDir_save, fn_prefix='Granada_')
df_result.head()

Top collaborators with Canada, overall:

df_can_col = df_pubs2.loc[df_pubs2['collaborators_of_interest'] != '{}']
print(len(df_can_col))
df_can_col.head()

df_result = df_can_col[['first_author', 'first_auth_inst', 'title', 'publication_date', 'cited_by_count', 'collaborators_of_interest', 'first_auth_orcid', 'id']]
df_result = df_result.sort_values(['cited_by_count', 'publication_date'], ascending=[False, False]).reset_index(drop=True)
df_result = df_result.drop_duplicates(subset=['id'], keep='first')
print(str(len(df_result)) + ' unique papers retrieved.')
df_result = df_result.rename(columns={
    'first_author': 'Spanish Researcher',
    'first_auth_inst': 'Institution',
    'title': 'Publication Title',
    'publication_date': 'Publication Date',
    'cited_by_count': 'Citation Count',
    'collaborators_of_interest': 'Canadian Collaborators',
    'first_auth_orcid': 'Academic Profile (ORCID)',
    'id': 'Publication Profile'
})
df_result = df_result.replace(np.nan, '')

#Process Canadian Collaborators
for i in range(0, len(df_result)):
    collab_dict_ = df_result['Canadian Collaborators'].iloc[i]
    if collab_dict_ != 'N':
        collab_dict = ast.literal_eval(collab_dict_)
        if len(collab_dict) > 0:
            collab_str = ''
            for auth in collab_dict:
                if collab_str != '':
                    collab_str += ', '
                collab_str += auth + ' (' + collab_dict[auth]['col_auth_inst']
                if collab_dict[auth]['col_auth_orcid'] == None:
                    collab_str += ')'
                else: 
                    collab_str += ', ' + collab_dict[auth]['col_auth_orcid'] + ')'
            df_result.iloc[i,5] = collab_str

df_result.head()

sheet_name_ = 'top-pubs with Canadian collab'
writer = pd.ExcelWriter(dataDir_save + 'Granada_works_can_col.xlsx')
df_result.to_excel(writer, sheet_name=sheet_name_, index=False)
# Auto-adjust columns' width
for column in df_result:
    if column == 'Publication Title':
        column_width = 100
    elif column == 'Canadian Collaborators':
        column_width = 75
    else:
        column_width = max(df_result[column].astype(str).map(len).max(), len(column))
    col_idx = df_result.columns.get_loc(column)
    writer.sheets[sheet_name_].set_column(col_idx, col_idx, column_width)

writer.save()

df_top_ac = df_can_col[['cited_by_count', 'publications', 'first_author']].groupby(['first_author']).sum()
df_top_ac = df_top_ac.sort_values('cited_by_count', ascending=False)
df_top_ac2 = df_top_ac.merge(df_pubs2[['first_author', 'first_auth_inst', 'first_auth_orcid']], how='left', on='first_author')
df_top_ac2 = df_top_ac2.rename(columns={
    'cited_by_count': 'by citations',
    'publications': 'by publications',
    'first_auth_inst': 'institution',
    'first_auth_orcid': 'orcid'
})
df_top_ac2 = df_top_ac2.drop_duplicates(subset=['orcid'])
df_top_ac2 = df_top_ac2.set_index('first_author')
df_top_ac2.index.names = ['Top-10 academics in Andalucía that collaborate with Canadian researchers']
df_top_ac2.head(10)
