==> 09_Prepare Data Model for App.ipynb <==
# Prepare Data Model for Application

#### Updated: Sep 28, 2022

# 

Prepare and strip down data model for potential Shiny or Power BI application.

import pandas as pd
import mysql.connector
import sqlalchemy as sa
from sqlalchemy import create_engine
import pymysql
import time

dataDir = '/Users/rnaidoo/Documents/Canada-Secure/GAC/2022_MDRID/Projects_data/OpenAlex/'

#  

#### Strip down data tables to necessary columns:

country_code = 'CA'
country_code_collab = 'ES'

df_pubs_first_auth = pd.read_csv(dataDir + 'works_' + country_code + '_first_auth/' + country_code_collab + '_col/' + 'works_' + country_code + '_first_auth_' + country_code_collab + '_col_since2017.csv')
print(len(df_pubs_first_auth))
df_pubs_first_auth.head()

df_pubs_first_auth2 = df_pubs_first_auth[['id', 'title', 'publication_date', 'authorships', 'cited_by_count', 'first_author', 'first_auth_orcid', 'first_auth_inst', 'first_auth_country', 'collaborators_of_interest']]
df_pubs_first_auth2.head()

df_pubs_first_auth2.to_csv(dataDir + 'works_' + country_code + '_first_auth/' + country_code_collab + '_col/' + 'works_' + country_code + '_first_auth_' + country_code_collab + '_col_since2017_APP.csv', index=False)

df_conc_first_auth = pd.read_csv(dataDir + 'works_' + country_code + '_first_auth/' + 'concepts_by_pub_' + country_code + '_first_auth_since2017.csv')
df_conc_first_auth

df_conc_first_auth2 = df_conc_first_auth[['pub_id', 'concept', 'concept_level']]
df_conc_first_auth2.to_csv(dataDir + 'works_' + country_code + '_first_auth/' + 'concepts_by_pub_' + country_code + '_first_auth_since2017_APP.csv', index=False)

#  

#### Think of how to leverage OpenAlex API calls, and SQL queries on our own database, as much as possible:

- Get data and process in batches...

- For a country overall, to produce tables of top institutions and top researchers (which would require processing of the entire 'works' table for a particular country), such tables could be pre-processed and loaded.

- To get works by region, select for 'works' table by 'first_auth_inst' column

- For analyzing the top collaborations, select works where 'collaborators_of_interest' is != '{}'
    - Similarly for the 'inverse works' table
    
- When searching by keyword, 
    - for concept search: on the 'concepts_by_pub' table select for rows where 'concept' column string matches with the keyword
    - for title search: on the 'works' table select for rows where 'title' column string matches with the keyword

#  

#### Upload data to MySQL database:

Connect to database with MySQL Connector

mydb = mysql.connector.connect(
    host="mdrid-1.cfgquvzxu6rh.ca-central-1.rds.amazonaws.com",
    user="admin",
    password="7jLkMAyKgESupoJiueft",
    database='MDRID_1'
)

print(mydb)

cursor = mydb.cursor()

Create table:

cursor.execute("CREATE TABLE concepts_by_pub_ES_first_auth_since2017_APP (\
               pub_id VARCHAR(255), \
               concept VARCHAR(255), \
               concept_level INTEGER\
               )")

cursor.execute("SHOW TABLES")
for x in cursor:
  print(x)

#    

Connect to database with SQLAlchemy:

server_name = 'mdrid-1.cfgquvzxu6rh.ca-central-1.rds.amazonaws.com'
database_name = 'MDRID_1'
username_ = 'admin'
pwd = '7jLkMAyKgESupoJiueft'

connection_url = sa.engine.URL.create(
    "mysql+pymysql",
    host=server_name,
    username=username_,
    password=pwd,
    database=database_name
)
engine = create_engine(connection_url)

Create table:

table_name = 'concepts_by_pub_' + country_code + '_first_auth_since2017_APP'

metadata_obj = sa.MetaData()
table = sa.Table(
    table_name,                                        
    metadata_obj,                                    
    sa.Column('pub_id', sa.types.NVARCHAR(length=255), primary_key=True),  
    sa.Column('concept', sa.types.NVARCHAR(length=255)),                    
    sa.Column('concept_level', sa.Integer),                
)
  
metadata_obj.create_all(engine)

df_load = pd.read_csv(dataDir + 'works_' + country_code + '_first_auth/' + table_name + '.csv')
print(len(df_load))
df_load.head()

#Get load time estimate (for very large loads)
'''df_test = df_load.head(10000)
start = time.time()
df_test.to_sql(name=table_name, con=engine, if_exists='replace', index=False)
end = time.time()
t = end - start
rate = t/len(df_test)
print('Load time: ' + str(round(t, 3)) + ' s (' + str(round(rate, 3)) + ' s/row)')
load_time_est = (len(df_load))*rate
print('Estimated load time for full table: ' + str(round(load_time_est/3600, 2)) + ' h...')'''

start = time.time()

df_load.to_sql(name=table_name, con=engine, if_exists='replace', index=False)

end = time.time()
t = end - start
rate = t/len(df_load)
print('Upload time: ' + str(round(t, 3)) + ' s (' + str(round(rate, 3)) + ' s/row)')

#  

Test load of the table back:

start = time.time()

df_loaded = pd.read_sql_query('SELECT * FROM ' + table_name, engine)

end = time.time()
print(len(df_loaded))
t = end - start
rate = t/len(df_loaded)
print('Download time: ' + str(round(t, 3)) + ' s (' + str(round(rate, 3)) + ' s/row)')
df_loaded.head()
