==> 11_Top Concepts In Publications Between Canada and Another Country.ipynb <==
# Top Concepts in Publications Between Canada and Another Country

#### Updated: Oct 27, 2022

#  

In response to BII and others, try to generate a list to show the top concepts for academic publications that Canadian and Spanish researchers collaborated on. 

import pandas as pd

dataDir = '/Users/rnaidoo/Documents/Canada-Secure/GAC/2022_MDRID/Projects_data/OpenAlex/'

#  

Load data:

country_code = 'ES'
country_code_collab = 'CA'

#Set-up directories
dataDir_first_auth = dataDir + 'works_' + country_code + '_first_auth/'
dataDir_col = dataDir_first_auth + country_code_collab + '_col/'
#Inverse
dataDir_first_auth_inv = dataDir + 'works_' + country_code_collab + '_first_auth/'
dataDir_col_inv = dataDir_first_auth_inv + country_code + '_col/'

df_pubs = pd.read_csv(dataDir_col + 'works_' + country_code + '_first_auth_' + country_code_collab + '_col_since2017.csv')
print(len(df_pubs))
df_pubs.head()

df_pubs_inv = pd.read_csv(dataDir_col_inv + 'works_' + country_code_collab + '_first_auth_' + country_code + '_col_since2017.csv')
print(len(df_pubs_inv))
df_pubs_inv.head()

df_conc = pd.read_csv(dataDir_first_auth + 'concepts_by_pub_' + country_code + '_first_auth_since2017.csv')
print(len(df_conc))
df_conc.head()

df_conc_inv = pd.read_csv(dataDir_first_auth_inv + 'concepts_by_pub_' + country_code_collab + '_first_auth_since2017.csv')
print(len(df_conc_inv))
df_conc_inv.head()

Isolate publications that have collaborators of interest:

df_pubs2 = df_pubs.loc[df_pubs['collaborators_of_interest'] != '{}']
print(len(df_pubs2))
df_pubs2.head()

df_pubs2_inv = df_pubs_inv.loc[df_pubs_inv['collaborators_of_interest'] != '{}']
print(len(df_pubs2_inv))
df_pubs2_inv.head()

Isolate by concept level 0:

concept_level = 0

df_conc2 = df_conc.loc[df_conc['concept_level'] == concept_level]
print(len(df_conc2))

df_conc2_inv = df_conc_inv.loc[df_conc_inv['concept_level'] == concept_level]
print(len(df_conc2_inv))

Merge concepts with publications:

df_pubs3 = df_pubs2.merge(df_conc2, how='right', left_on='id', right_on='pub_id')
df_pubs3 = df_pubs3[df_pubs3['id'].notnull()]
print(len(df_pubs3))
df_pubs3.head()

df_pubs3_inv = df_pubs2_inv.merge(df_conc2_inv, how='right', left_on='id', right_on='pub_id')
df_pubs3_inv = df_pubs3_inv[df_pubs3_inv['id'].notnull()]
print(len(df_pubs3_inv))
df_pubs3_inv.head()

df_pubs4 = pd.concat([df_pubs3, df_pubs3_inv])
print(len(df_pubs4))
df_pubs4.head()

Group by concept:

df_pubs5 = df_pubs4[['concept', 'cited_by_count']]
df_pubs5['publications'] = 1
print(len(df_pubs5))
df_pubs5.head()

df_result = df_pubs5.groupby(['concept']).sum().sort_values('cited_by_count', ascending=False)
df_result

df_result.to_csv(dataDir_col + 'works_by_concept_' + str(concept_level) + '_' + country_code + '-' + country_code_collab + '.csv')

#  

Group by concept level 1:

concept_level = 3

df_conc2 = df_conc.loc[df_conc['concept_level'] == concept_level]
print(len(df_conc2))
df_pubs3 = df_pubs2.merge(df_conc2, how='right', left_on='id', right_on='pub_id')
df_pubs3 = df_pubs3[df_pubs3['id'].notnull()]
print(len(df_pubs3))
df_conc2_inv = df_conc_inv.loc[df_conc_inv['concept_level'] == concept_level]
print(len(df_conc2_inv))
df_pubs3_inv = df_pubs2_inv.merge(df_conc2_inv, how='right', left_on='id', right_on='pub_id')
df_pubs3_inv = df_pubs3_inv[df_pubs3_inv['id'].notnull()]
print(len(df_pubs3_inv))

df_pubs4 = pd.concat([df_pubs3, df_pubs3_inv])
print(len(df_pubs4))

df_pubs5 = df_pubs4[['concept', 'cited_by_count']]
df_pubs5['publications'] = 1

df_result = df_pubs5.groupby(['concept']).sum().sort_values('cited_by_count', ascending=False)
df_result.to_csv(dataDir_col + 'works_by_concept_' + str(concept_level) + '_' + country_code + '-' + country_code_collab + '.csv')
df_result
