==> 12_Regional Profile III.ipynb <==
# Regional Profile III

#### Updated: Aug 19, 2023

#  

Find the top collaborations between Canada and the top-30 Spanish research instiutions.

Ref: https://www.scimagoir.com/rankings.php?sector=Higher+educ.&country=ESP&ranking=Research

import pandas as pd
import numpy as np
import ast
import os
import requests
import xmltodict
import datetime
from fuzzywuzzy import fuzz
from fuzzywuzzy import process
from deep_translator import GoogleTranslator

import warnings
warnings.filterwarnings('ignore')

dataDir = '/Users/rnaidoo/Documents/Canada-Secure/GAC/2022_MDRID/Projects_data/OpenAlex/'

#from translate import Translator
#translator = Translator(from_lang="spanish",to_lang="english")

#  

#### Functions:

Define region:

def search_institutions(df_pubs, inst_search_terms=[]):
    
    if len(inst_search_terms) > 0 : 
        inst_search_results = []
        for search_term in inst_search_terms:
            inst_search_results = inst_search_results + list(df_pubs.loc[df_pubs['first_auth_inst'].str.contains(search_term)]['first_auth_inst'].unique())
        df_insts = pd.DataFrame({'inst_search': inst_search_results})
        df_pubs2 = df_pubs.merge(df_insts, how='inner', left_on='first_auth_inst', right_on='inst_search')
        
    else:
        inst_search_results = list(df_pubs['first_auth_inst'].unique())
        df_insts = pd.DataFrame({'inst_search': inst_search_results})
        df_pubs2 = df_pubs
        
    return [df_insts, df_pubs2]

Summarize region:

def top_institutions(df_pubs, reg_name, top=100, search_query=''):
    
    df_pubs['cited_by_count'] = pd.to_numeric(df_pubs['cited_by_count'])
    df_pubs['publications'] = 1
    
    df_top_inst = df_pubs[['cited_by_count', 'publications', 'first_auth_inst']].groupby(['first_auth_inst']).sum()
    df_top_inst = df_top_inst.sort_values('cited_by_count', ascending=False)
    df_top_inst = df_top_inst.replace(np.nan, '')
    df_top_inst = df_top_inst.rename(columns={
        'cited_by_count': 'by citations',
        'publications': 'by publications'
    })
    if len(df_top_inst) < top:
        top = len(df_top_inst)
    table_title = 'Top-' + str(top) + ' research institutions in ' + reg_name
    if len(search_query) > 0:
        table_title = table_title + " for '" + search_query + "'"
    df_top_inst.index.names = [table_title]
    
    return df_top_inst.head(top)

def top_researchers(df_pubs, reg_name, top=100, search_query=''):
    
    df_pubs['cited_by_count'] = pd.to_numeric(df_pubs['cited_by_count'])
    df_pubs['publications'] = 1
    
    df_top_ac = df_pubs[['cited_by_count', 'publications', 'first_author']].groupby(['first_author']).sum()
    df_top_ac = df_top_ac.sort_values('cited_by_count', ascending=False)
    df_top_ac2 = df_top_ac.merge(df_pubs[['first_author', 'first_auth_inst', 'first_auth_orcid']], how='left', on='first_author')
    df_top_ac2 = df_top_ac2.replace(np.nan, '')
    df_top_ac2 = df_top_ac2.rename(columns={
        'cited_by_count': 'by citations',
        'publications': 'by publications',
        'first_auth_inst': 'institution',
        'first_auth_orcid': 'orcid'
    })
    df_top_ac2 = df_top_ac2.drop_duplicates(subset=['orcid'])
    df_top_ac2 = df_top_ac2.drop_duplicates(subset=['first_author'])
    df_top_ac2 = df_top_ac2.set_index('first_author')
    
    if len(df_top_ac2) < top:
        top = len(df_top_ac2)
    table_title = 'Top-' + str(top) + ' academics in ' + reg_name
    
    if len(search_query) > 0:
        table_title = table_title + " for '" + search_query + "'"
    df_top_ac2.index.names = [table_title]
    
    df_top_ac2 = df_top_ac2.head(top)
    
    #Check ORCID API for latest role
    df_top_ac2['latest_org'] = pd.Series()
    df_top_ac2['latest_role'] = pd.Series()
    df_top_ac2['latest_role_start'] = pd.Series()
    for i in range(0, len(df_top_ac2)):
        ORCID = df_top_ac2['orcid'].iloc[i]
        ORCID, latest_org, latest_role, latest_role_start = get_latest_role_ORCID2(ORCID)
        df_top_ac2.loc[df_top_ac2.index[i], 'latest_org'] = latest_org
        df_top_ac2.loc[df_top_ac2.index[i], 'latest_role'] = latest_role
        df_top_ac2.loc[df_top_ac2.index[i], 'latest_role_start'] = latest_role_start
        
    return df_top_ac2

Explore collaborations:

def proc_collab(df_in):
    
    for i in range(0, len(df_in)):
        collab_dict_ = df_in['collaborators_of_interest'].iloc[i]
        if collab_dict_ != '{}':
            collab_dict = ast.literal_eval(collab_dict_)
            if len(collab_dict) > 0:
                collab_str = ''
                for auth in collab_dict:
                    if collab_str != '':
                        collab_str += ', '
                    collab_str += auth + ' (' + collab_dict[auth]['col_auth_inst']
                    if collab_dict[auth]['col_auth_orcid'] == None:
                        collab_str += ')'
                    else: 
                        collab_str += ', ' + collab_dict[auth]['col_auth_orcid'] + ')'
                df_in['collaborators_of_interest'].iloc[i] = collab_str
    
    return df_in

def proc_collab_inv(df_in, reg_insts):
    for i in range(0, len(df_in)):
        collab_dict_ = df_in['collaborators_of_interest'].iloc[i]
        if collab_dict_ != '{}':
            collab_dict = ast.literal_eval(collab_dict_)
            if len(collab_dict) > 0:
                collab_str = ''
                for auth in collab_dict:
                    for reg_inst in reg_insts:
                        if collab_dict[auth]['col_auth_inst'] == reg_inst:
                            if collab_str != '':
                                collab_str += ', '
                            collab_str += auth + ' (' + collab_dict[auth]['col_auth_inst']
                            if collab_dict[auth]['col_auth_orcid'] == None:
                                collab_str += ')'
                            else: 
                                collab_str += ', ' + collab_dict[auth]['col_auth_orcid'] + ')'
                if collab_str == '':
                    collab_str = '{}'
                df_in['collaborators_of_interest'].iloc[i] = collab_str
    
    return df_in

def write_pubs_by_top_collab(df_result, reg_name, country_code_collab, dataDir_save, inv=False):
    
    df_result = df_result.replace(np.nan, '')
    df_result['collaborators_of_interest'] = df_result['collaborators_of_interest'].replace('{}', 'N')
    
    #Check ORCID API for latest role
    df_result['latest_org'] = pd.Series()
    df_result['latest_role'] = pd.Series()
    df_result['latest_role_start'] = pd.Series()
    for i in range(0, len(df_result)):
        ORCID = df_result['first_auth_orcid'].iloc[i]
        ORCID, latest_org, latest_role, latest_role_start = get_latest_role_ORCID2(ORCID)
        df_result.loc[df_result.index[i], 'latest_org'] = latest_org
        df_result.loc[df_result.index[i], 'latest_role'] = latest_role
        df_result.loc[df_result.index[i], 'latest_role_start'] = latest_role_start
    
    df_result = df_result.rename(columns={
        'first_auth_inst': 'Institution',
        'title': 'Publication Title',
        'publication_date': 'Publication Date',
        'cited_by_count': 'Citation Count',
        'first_auth_orcid': 'Academic Profile (ORCID)',
        'id': 'Publication Profile'
    })
    if inv:
        sheet_name_ = 'top-pubs with ' + reg_name + ' collab'
        sheet_name_ = sheet_name_[:30]
        writer = pd.ExcelWriter(dataDir_save + country_code_collab + '_works_' + reg_name + '_col.xlsx')
        df_result = df_result.rename(columns={
            'first_author': country_code_collab + ' Researcher',
            'collaborators_of_interest': reg_name + ' Collaborators'
        })
    else:        
        sheet_name_ = 'top-pubs with ' + country_code_collab + ' collab'
        sheet_name_ = sheet_name_[:30]
        writer = pd.ExcelWriter(dataDir_save + reg_name + '_works_' + country_code_collab + '_col.xlsx')
        df_result = df_result.rename(columns={
            'first_author': reg_name + ' Researcher',
            'collaborators_of_interest': country_code_collab + ' Collaborators'
        })    
    df_result.to_excel(writer, sheet_name=sheet_name_, index=False)
    
    # Auto-adjust columns' width
    for column in df_result.columns:
        if column == 'Publication Title':
            column_width = 100
        elif column == country_code_collab + ' Collaborators':
            column_width = 75
        elif column == reg_name + ' Collaborators':
            column_width = 75
        else:
            column_width = max(df_result[column].astype(str).map(len).max(), len(column))
        col_idx = df_result.columns.get_loc(column)
        writer.sheets[sheet_name_].set_column(col_idx, col_idx, column_width)

    writer.save()

def top_collab_summary_table(df_pubs, reg_name, country_code_collab, top=100, inv=False, search_query=''):
    
    df_col = df_pubs.loc[df_pubs['collaborators_of_interest'] != '{}']
    df_col['cited_by_count'] = pd.to_numeric(df_col['cited_by_count'])
    df_col['publications'] = 1
    
    df_top_ac = df_col[['cited_by_count', 'publications', 'first_author']].groupby(['first_author']).sum()
    df_top_ac = df_top_ac.sort_values('cited_by_count', ascending=False)
    df_top_ac2 = df_top_ac.merge(df_pubs[['first_author', 'first_auth_inst', 'first_auth_orcid']], how='left', on='first_author')
    df_top_ac2 = df_top_ac2.replace(np.nan, '')
    df_top_ac2 = df_top_ac2.rename(columns={
        'cited_by_count': 'by citations',
        'publications': 'by publications',
        'first_auth_inst': 'institution',
        'first_auth_orcid': 'orcid'
    })
    df_top_ac2 = df_top_ac2.drop_duplicates(subset=['orcid'])
    df_top_ac2 = df_top_ac2.drop_duplicates(subset=['first_author'])
    df_top_ac2 = df_top_ac2.set_index('first_author')
    if len(df_top_ac2) < top:
        top = len(df_top_ac2)
    
    if inv:
        table_title = 'Top-' + str(top) + ' academics in ' + country_code_collab + ' that collaborated with ' + reg_name + ' researchers'
        if len(search_query) > 0:
            table_title = table_title + ", on \'" + search_query + "'"
        df_top_ac2.index.names = [table_title]        
    else:
        table_title = 'Top-' + str(top) + ' academics in ' + reg_name + ' that collaborated with ' + country_code_collab + ' researchers'
        if len(search_query) > 0:
            table_title = table_title + ", on \'" + search_query + "'"
        df_top_ac2.index.names = [table_title]
    
    df_top_ac2 = df_top_ac2.head(top)
    
    #Check ORCID API for latest role
    df_top_ac2['latest_org'] = pd.Series()
    df_top_ac2['latest_role'] = pd.Series()
    df_top_ac2['latest_role_start'] = pd.Series()
    for i in range(0, len(df_top_ac2)):
        ORCID = df_top_ac2['orcid'].iloc[i]
        ORCID, latest_org, latest_role, latest_role_start = get_latest_role_ORCID2(ORCID)
        df_top_ac2.loc[df_top_ac2.index[i], 'latest_org'] = latest_org
        df_top_ac2.loc[df_top_ac2.index[i], 'latest_role'] = latest_role
        df_top_ac2.loc[df_top_ac2.index[i], 'latest_role_start'] = latest_role_start
    
    return df_top_ac2

def top_collab(df_pubs, reg_name, country_code_collab, dataDir_save, top=100):
    
    df_col = df_pubs.loc[df_pubs['collaborators_of_interest'] != '{}']
    
    df_result = df_col[['first_author', 'first_auth_inst', 'title', 'publication_date', 'cited_by_count', 'collaborators_of_interest', 'first_auth_orcid', 'id']]
    df_result = df_result.sort_values(['cited_by_count', 'publication_date'], ascending=[False, False]).reset_index(drop=True)
    df_result = df_result.drop_duplicates(subset=['id'], keep='first')
    print(str(len(df_result)) + ' unique papers retrieved.')
    
    #Process collaborators of interest
    df_result = proc_collab(df_in=df_result)
    
    #Write results to Excel spreadsheet
    write_pubs_by_top_collab(df_result=df_result, reg_name=reg_name, country_code_collab=country_code_collab, dataDir_save=dataDir_save)
    
    #Produce summary table
    df_top = top_collab_summary_table(df_pubs=df_result, reg_name=reg_name, country_code_collab=country_code_collab, top=top, inv=False)
    
    return df_result, df_top

def top_collab_inv(df_pubs, reg_name, reg_insts, country_code_collab, dataDir_save, top=100): #reg_insts from df_insts
    
    df_col = df_pubs.loc[df_pubs['collaborators_of_interest'] != '{}']
    
    df_result = df_col[['first_author', 'first_auth_inst', 'title', 'publication_date', 'cited_by_count', 'collaborators_of_interest', 'first_auth_orcid', 'id']]
    df_result = df_result.sort_values(['cited_by_count', 'publication_date'], ascending=[False, False]).reset_index(drop=True)
    df_result = df_result.drop_duplicates(subset=['id'], keep='first')
    
    #Find the papers where researchers in the region of interest are collaborators
    reg_col_indicies = []
    for i in range(0, len(df_result)):
        collab_dict_ = df_result['collaborators_of_interest'].iloc[i]
        collab_dict = ast.literal_eval(collab_dict_)
        if len(collab_dict) > 0:
            for auth in collab_dict:
                for reg_inst in reg_insts:
                    if collab_dict[auth]['col_auth_inst'] == reg_inst:
                        reg_col_indicies.append(i) 
    df_result = df_result.iloc[reg_col_indicies]
    print(str(len(df_result)) + ' unique papers retrieved.')
    
    #Process collaborators of interest
    df_result = proc_collab_inv(df_in=df_result, reg_insts=reg_insts)
    
    #Write results to Excel spreadsheet
    write_pubs_by_top_collab(df_result=df_result, reg_name=reg_name, country_code_collab=country_code_collab, dataDir_save=dataDir_save, inv=True)    
    
    #Produce summary table
    df_top = top_collab_summary_table(df_pubs=df_result, reg_name=reg_name, country_code_collab=country_code_collab, top=top, inv=True)
    
    return df_result, df_top

def top_collab_pairs_table(df_search, df_search_inv, top=100):
    
    df_col = df_search.loc[df_search['collaborators_of_interest'] != '{}']
    df_col_inv = df_search_inv.loc[df_search_inv['collaborators_of_interest'] != '{}']
    df_col_tot = pd.concat([df_col, df_col_inv])
    df_top_pairs = pd.DataFrame() #pd.DataFrame(columns=['First Author', 'Collaborator', 'Citations', 'Concept', 'Publication Profile'])
    if len(df_col_tot) > 0:
        if 'concept' in df_col_tot.columns:
            df_col_tot = df_col_tot[['first_author', 'first_auth_inst', 'cited_by_count', 'collaborators_of_interest', 'concept', 'id']] #'title', 'publication_date'
        else:
            df_col_tot = df_col_tot[['first_author', 'first_auth_inst', 'cited_by_count', 'collaborators_of_interest', 'id']] #'title', 'publication_date',
        df_col_tot = df_col_tot.sort_values(['cited_by_count'], ascending=[False]).reset_index(drop=True) #'publication_date'
        df_col_tot = df_col_tot.drop_duplicates(subset=['id'], keep='first')

        if len(df_col_tot) < top:
            top = len(df_col_tot)
        for i in range(0, int(top)):
            first_auth_str = df_col_tot['first_author'].iloc[i] + ' (' + df_col_tot['first_auth_inst'].iloc[i] + ')'
            col_str = df_col_tot['collaborators_of_interest'].iloc[i].split(',')[0]
            if ',' in df_col_tot['collaborators_of_interest'].iloc[i]:
                col_str += ')'
            row = pd.DataFrame({
                  'First Author': first_auth_str,
                  'Collaborator': col_str,
                  'Citations': df_col_tot['cited_by_count'].iloc[i],
                  'Concept': df_col_tot['concept'].iloc[i],
                  #'Publication Title': df_col_tot['title'].iloc[i],
                  #'Date': df_col_tot['publication_date'].iloc[i],
                  'Publication Profile': df_col_tot['id'].iloc[i]
              }, index=[i])
            df_top_pairs = pd.concat([df_top_pairs, row], axis=0, ignore_index=True)
    
    return df_top_pairs

def top_collab_pairs_table2(country_code, df_search, country_code_collab, df_search_inv, top=100):
    
    df_col = df_search.loc[df_search['collaborators_of_interest'] != '{}']
    if len(df_col) > 0:
        if 'concept' in df_col.columns:
            df_col = df_col[['first_author', 'first_auth_inst', 'cited_by_count', 'collaborators_of_interest', 'concept', 'id']] #'title', 'publication_date'
        else:
            df_col = df_col[['first_author', 'first_auth_inst', 'cited_by_count', 'collaborators_of_interest', 'id']] #'title', 'publication_date',
        df_col = df_col.sort_values(['cited_by_count'], ascending=[False]).reset_index(drop=True) #'publication_date'
        df_col = df_col.drop_duplicates(subset=['id'], keep='first')
    df_col = df_col.head(int(top))
    df_col[country_code + ' Author'] = ''
    df_col[country_code_collab + ' Author'] = ''
    df_col_top_pairs = pd.DataFrame()
    for i in range(0, len(df_col)):
        first_auth_str = df_col['first_author'].iloc[i] + ' (' + df_col['first_auth_inst'].iloc[i] + ')*'
        col_str = df_col['collaborators_of_interest'].iloc[i].split(',')[0]
        if ',' in df_col['collaborators_of_interest'].iloc[i]:
                col_str += ')'
        if 'concept' in df_col.columns:
            row = pd.DataFrame({
                    country_code + ' Author': first_auth_str,
                    country_code_collab + ' Author': col_str,
                    'Citations': df_col['cited_by_count'].iloc[i],
                    'Concept': df_col['concept'].iloc[i],
                    'Publication Profile': df_col['id'].iloc[i]
                }, index=[i])
            row = row.fillna('')
        else:
            row = pd.DataFrame({
                    country_code + ' Author': first_auth_str,
                    country_code_collab + ' Author': col_str,
                    'Citations': df_col['cited_by_count'].iloc[i],
                    'Publication Profile': df_col['id'].iloc[i]
                }, index=[i])
        df_col_top_pairs = pd.concat([df_col_top_pairs, row], axis=0, ignore_index=True)

    #Inverse
    df_col_inv = df_search_inv.loc[df_search_inv['collaborators_of_interest'] != '{}']
    if len(df_col_inv) > 0:
        if 'concept' in df_col_inv.columns:
            df_col_inv = df_col_inv[['first_author', 'first_auth_inst', 'cited_by_count', 'collaborators_of_interest', 'concept', 'id']] #'title', 'publication_date'
        else:
            df_col_inv = df_col_inv[['first_author', 'first_auth_inst', 'cited_by_count', 'collaborators_of_interest', 'id']] #'title', 'publication_date',
        df_col_inv = df_col_inv.sort_values(['cited_by_count'], ascending=[False]).reset_index(drop=True) #'publication_date'
        df_col_inv = df_col_inv.drop_duplicates(subset=['id'], keep='first')
    df_col_inv = df_col_inv.head(int(top))
    df_col_inv[country_code + ' Author'] = ''
    df_col_inv[country_code_collab + ' Author'] = ''
    df_col_inv_top_pairs = pd.DataFrame()
    for i in range(0, len(df_col_inv)):
        first_auth_str = df_col_inv['first_author'].iloc[i] + ' (' + df_col_inv['first_auth_inst'].iloc[i] + ')*'
        col_str = df_col_inv['collaborators_of_interest'].iloc[i].split(',')[0]
        if ',' in df_col_inv['collaborators_of_interest'].iloc[i]:
                col_str += ')'
        if 'concept' in df_col_inv.columns:
            row = pd.DataFrame({
                    country_code + ' Author': col_str,
                    country_code_collab + ' Author': first_auth_str,
                    'Citations': df_col_inv['cited_by_count'].iloc[i],
                    'Concept': df_col_inv['concept'].iloc[i],
                    'Publication Profile': df_col_inv['id'].iloc[i]
                }, index=[i])
        else:
            row = pd.DataFrame({
                    country_code + ' Author': col_str,
                    country_code_collab + ' Author': first_auth_str,
                    'Citations': df_col_inv['cited_by_count'].iloc[i],
                    'Publication Profile': df_col_inv['id'].iloc[i]
                }, index=[i])
        df_col_inv_top_pairs = pd.concat([df_col_inv_top_pairs, row], axis=0, ignore_index=True)
    
    df_top_pairs = pd.concat([df_col_top_pairs, df_col_inv_top_pairs])
    if len(df_top_pairs) > 0:
        df_top_pairs = df_top_pairs.sort_values(['Citations'], ascending=[False]).reset_index(drop=True)
    df_top_pairs = df_top_pairs.head(int(top))
    
    return df_top_pairs

def write_collab_summary(df_col_pairs, df_top_col_A, df_top_col_B, df_A, df_A_inst, reg_name, country_code_collab, dataDir_save):
    
    writer = pd.ExcelWriter(dataDir_save + 'Summary-' + reg_name + '-' + country_code_collab + '.xlsx')
    
    tables = [df_col_pairs, df_top_col_A, df_top_col_B, df_A, df_A_inst]
    sheet_names = ['Top collaborations_pairs', 'Top acad col - A', 'Top acad col - B', 'Top academics - A', 'Top institutions - A']
    
    for i in range(0, len(tables)):
        sheet_name_ = sheet_names[i]
        df_save = tables[i]
        if i > 0:
            df_save = df_save.reset_index()
        df_save.to_excel(writer, sheet_name=sheet_name_, index=False)
        
        #Set column widths
        for column in df_save.columns:
            #print(column)
            column_width = max(df_save[column].astype(str).map(len).max(), len(column))
            if column_width > 100:
                column_width = 100
            col_idx = df_save.columns.get_loc(column)
            writer.sheets[sheet_name_].set_column(col_idx, col_idx, column_width)
            
    writer.save()
    

ORCID:

#Search ORCID API, by table of works:
def get_latest_role_ORCID(df_result, orcid_col_name):
    
    #Get ORCID 'read-public' access token:
    url = 'https://pub.orcid.org/oauth/token'
    headers = {
      'Accept': 'application/json'
    }
    data = {
      'client_id': 'APP-2K656AL70DN662NJ',
      'client_secret': '7770448b-d0aa-4b2c-8f2d-44d339cf1532',
      'grant_type': 'client_credentials',
      'scope': '/read-public'
    }
    response = requests.post(url=url, data=data, headers=headers)
    
    if response.status_code == 200:
        token = response.json()['access_token']
    
        df_result['Last_known_org'] = ''
        df_result['Last_known_role'] = ''
        df_result['Last_known_role_start_date'] = ''
        print(str(len(df_result)) + ' ORCID profiles to check:')
        for i in range(0, len(df_result)):
            if i%100 == 0:
                print('checking ' + str(i) + 'th ORCID profile')
            ORCID_ = df_result[orcid_col_name].iloc[i]
            if ORCID_ != None:
                ORCID_ = str(ORCID_)
                if 'https' in ORCID_:
                    ORCID = ORCID_.replace('https://orcid.org/', '')
                    url = 'https://pub.orcid.org/v3.0/' + ORCID + '/employments'
                    headers = {
                      'Accept': 'application/vnd.orcid+xml',
                      'Authorization': token
                    }
                    data = {
                      'Bearer': token
                    }
                    response = requests.get(url=url, data=data, headers=headers)
                    if response.status_code == 200:
                        data_dict = xmltodict.parse(response.content)
                        try:
                            latest_employment_summary = data_dict['activities:employments']['activities:affiliation-group'][0]['employment:employment-summary']
                            try:
                                latest_role = latest_employment_summary['common:role-title']
                            except:
                                latest_role = ''
                            try:
                                latest_role_start_year = latest_employment_summary['common:start-date']['common:year']
                                latest_role_start_month = latest_employment_summary['common:start-date']['common:month']
                                latest_role_start_day = latest_employment_summary['common:start-date']['common:day']
                                latest_role_start = datetime.date(int(latest_role_start_year), int(latest_role_start_month), int(latest_role_start_day))
                            except:
                                latest_role_start = ''
                            try:
                                latest_org = latest_employment_summary['common:organization']['common:name']
                            except:
                                latest_org = ''
                        except:
                            continue

                        #Write to df_result:
                        df_result.at[i, 'Last_known_org'] = latest_org
                        df_result.at[i, 'Last_known_role'] = latest_role
                        df_result.at[i, 'Last_known_role_start_date'] = latest_role_start
                        '''if 'concept' in df_result.columns:
                            df_result = df_result[['first_author', 'first_auth_inst', 'concept', 'title', 'publication_date', 'cited_by_count', 
                                                'collaborators_of_interest', 'first_auth_orcid', 'id', 'Last_known_org', 'Last_known_role',
                                                'Last_known_role_start_date']]
                        else:
                            df_result = df_result[['first_author', 'first_auth_inst', 'title', 'publication_date', 'cited_by_count', 
                                               'collaborators_of_interest', 'first_auth_orcid', 'id', 'Last_known_org', 'Last_known_role',
                                               'Last_known_role_start_date']]'''
    
    else:
        print('Error: Could not connect to ORCID API')
    
    return df_result

#Search ORCID API, by ORCID
def get_latest_role_ORCID2(ORCID):
    
    #Get ORCID 'read-public' access token:
    url = 'https://pub.orcid.org/oauth/token'
    headers = {
      'Accept': 'application/json'
    }
    data = {
      'client_id': 'APP-2K656AL70DN662NJ',
      'client_secret': '7770448b-d0aa-4b2c-8f2d-44d339cf1532',
      'grant_type': 'client_credentials',
      'scope': '/read-public'
    }
    response = requests.post(url=url, data=data, headers=headers)
    
    latest_org = ''
    latest_role = ''
    latest_role_start = ''
    
    if response.status_code == 200:
        token = response.json()['access_token']
    
        if ORCID != None:
            ORCID = str(ORCID)
            if 'https' in ORCID:
                ORCID = ORCID.replace('https://orcid.org/', '')
                url = 'https://pub.orcid.org/v3.0/' + ORCID + '/employments'
                headers = {
                  'Accept': 'application/vnd.orcid+xml',
                  'Authorization': token
                }
                data = {
                  'Bearer': token
                }
                response = requests.get(url=url, data=data, headers=headers)
                if response.status_code == 200:
                    data_dict = xmltodict.parse(response.content)
                    try:
                        latest_employment_summary = data_dict['activities:employments']['activities:affiliation-group'][0]['employment:employment-summary']
                        try:
                            latest_role = latest_employment_summary['common:role-title']
                        except:
                            latest_role = ''
                        try:
                            latest_role_start_year = latest_employment_summary['common:start-date']['common:year']
                            latest_role_start_month = latest_employment_summary['common:start-date']['common:month']
                            latest_role_start_day = latest_employment_summary['common:start-date']['common:day']
                            latest_role_start = datetime.date(int(latest_role_start_year), int(latest_role_start_month), int(latest_role_start_day))
                        except:
                            latest_role_start = ''
                        try:
                            latest_org = latest_employment_summary['common:organization']['common:name']
                        except:
                            latest_org = ''
                    except:
                        latest_role = ''
    
    else:
        print('Error: Could not connect to ORCID API')
    
    return ORCID, latest_org, latest_role, latest_role_start

#  

#### Load requisite data:

country_code = 'ES'
country_code_collab = 'CA'
year_i = 2018

df_pubs_A_first_auth = pd.read_csv(dataDir + 'works_' + country_code + '_first_auth/' + country_code_collab + '_col/' + 'works_' + country_code + '_first_auth_' + country_code_collab + '_col_since' + str(year_i) +'.csv')
print(len(df_pubs_A_first_auth))
df_pubs_A_first_auth.head()

df_pubs_B_first_auth = pd.read_csv(dataDir + 'works_' + country_code_collab + '_first_auth/' + country_code + '_col/' + 'works_' + country_code_collab + '_first_auth_' + country_code + '_col_since' + str(year_i) + '.csv')
print(len(df_pubs_B_first_auth))
df_pubs_B_first_auth.head()

df_conc_A_first_auth = pd.read_csv(dataDir + 'works_' + country_code + '_first_auth/' + 'concepts_by_pub_' + country_code + '_first_auth_since' + str(year_i) + '.csv')
df_conc_A_first_auth

df_conc_B_first_auth = pd.read_csv(dataDir + 'works_' + country_code_collab + '_first_auth/' + 'concepts_by_pub_' + country_code_collab + '_first_auth_since' + str(year_i) + '.csv')
df_conc_B_first_auth

unq_insts = df_pubs_A_first_auth['first_auth_inst'].unique()
unq_insts

#  

#### Specify a region of interest (if necessary):

Search for institutions for a particular region:

inst_search_term_ES = 'Universidad Complutense de Madrid'

match_list = []
for i in range(0, len(unq_insts)):
    match_list.append(fuzz.partial_ratio(inst_search_term_ES, unq_insts[i]))
max_index = match_list.index(max(match_list))
inst_search_terms = []
inst_search_terms.append(unq_insts[max_index])

reg_name = inst_search_terms[0]

if reg_name != '':
    dataDir_reg = dataDir+'works_' + country_code + '_first_auth/' + country_code_collab + '_col/Regional/' + reg_name + '/'
    os.makedirs(dataDir_reg, exist_ok=True)
else:
    dataDir_reg = dataDir+'works_' + country_code + '_first_auth/' + country_code_collab + '_col/'
    reg_name = country_code

df_insts, df_pubs_reg = search_institutions(df_pubs=df_pubs_A_first_auth, inst_search_terms=inst_search_terms)
reg_insts = df_insts['inst_search'].unique()
reg_insts

ORCID check:

#df_pubs_reg = get_latest_role_ORCID(df_pubs_reg, 'first_auth_orcid')

Get tables:

df_A_inst = top_institutions(df_pubs=df_pubs_reg, reg_name=reg_name, top=10)
df_A_inst

df_A = top_researchers(df_pubs=df_pubs_reg, reg_name=reg_name, top=10)
df_A

df_col, df_top_col_A = top_collab(df_pubs=df_pubs_reg, reg_name=reg_name, country_code_collab=country_code_collab, dataDir_save=dataDir_reg, top=10)
df_top_col_A

df_col_inv, df_top_col_B = top_collab_inv(df_pubs=df_pubs_B_first_auth, reg_name=reg_name, reg_insts=reg_insts, country_code_collab=country_code_collab, dataDir_save=dataDir_reg, top=10)
df_top_col_B

df_col_pairs = top_collab_pairs_table2(country_code=country_code, df_search=df_col, country_code_collab=country_code_collab, df_search_inv=df_col_inv)
df_col_pairs

Write tables to summary sheet:

write_collab_summary(df_col_pairs=df_col_pairs, df_top_col_A=df_top_col_A, df_top_col_B=df_top_col_B, df_A=df_A, df_A_inst=df_A_inst, reg_name=reg_name, country_code_collab=country_code_collab, dataDir_save=dataDir_reg)

#  

#### Produce tables for the top-30 Spanish research institutions:

df_ES_insts = pd.read_csv(dataDir + 'works_' + country_code + '_first_auth/' + 'ScimagoIR 2022 - Research Rank - Universities - ESP.csv', sep=';')
df_ES_insts.head()

for i in range(0, len(df_ES_insts)):
    inst_search_term_ES = df_ES_insts['Institution'].iloc[i]
    inst_search_term_ES = inst_search_term_ES.replace(' *', '')
    print_str = inst_search_term_ES + ' ---> '
    
    #Translate from Spanish to English
    inst_search_term_ES_trans_EN = GoogleTranslator(source='spanish', target='english').translate(inst_search_term_ES)
    #inst_search_term_ES_trans_EN = translator.translate(inst_search_term_ES)
    #print(inst_search_term_ES_trans_EN)
    print_str += inst_search_term_ES_trans_EN + ' ---> '
    
    #Fuzzy match search for institution in OpenAlex data
    match_list = []
    for i in range(0, len(unq_insts)):
        match_list.append(fuzz.token_sort_ratio(inst_search_term_ES_trans_EN, unq_insts[i]))
    max_index = match_list.index(max(match_list))
    inst_search_terms = []
    inst_search_terms.append(unq_insts[max_index])
    
    reg_name = inst_search_terms[0]
    if reg_name != '':
        dataDir_reg = dataDir+'works_' + country_code + '_first_auth/' + country_code_collab + '_col/Regional/' + reg_name + '/'
        os.makedirs(dataDir_reg, exist_ok=True)
    else:
        dataDir_reg = dataDir+'works_' + country_code + '_first_auth/' + country_code_collab + '_col/'
        reg_name = country_code
    
    df_insts, df_pubs_reg = search_institutions(df_pubs=df_pubs_A_first_auth, inst_search_terms=inst_search_terms)
    reg_insts = df_insts['inst_search'].unique()
    
    for inst_search_terms in inst_search_terms:
        print_str += inst_search_terms + ', '
    print(print_str)
    
    #Get tables
    df_A_inst = top_institutions(df_pubs=df_pubs_reg, reg_name=reg_name, top=10)
    df_A = top_researchers(df_pubs=df_pubs_reg, reg_name=reg_name, top=10)
    df_col, df_top_col_A = top_collab(df_pubs=df_pubs_reg, reg_name=reg_name, country_code_collab=country_code_collab, dataDir_save=dataDir_reg, top=10)
    df_col_inv, df_top_col_B = top_collab_inv(df_pubs=df_pubs_B_first_auth, reg_name=reg_name, reg_insts=reg_insts, country_code_collab=country_code_collab, dataDir_save=dataDir_reg, top=10)
    df_col_pairs = top_collab_pairs_table2(country_code=country_code, df_search=df_col, country_code_collab=country_code_collab, df_search_inv=df_col_inv)

    #Write tables to summary sheet
    write_collab_summary(df_col_pairs=df_col_pairs, df_top_col_A=df_top_col_A, df_top_col_B=df_top_col_B, df_A=df_A, df_A_inst=df_A_inst, reg_name=reg_name, country_code_collab=country_code_collab, dataDir_save=dataDir_reg)
